refactor: change config files
