+++ .browserslistrc
@@ -1,0 +1,6 @@
+# run `npx browserslist` to see the current browserversions
+
+last 1 Chrome versions
+last 1 ChromeAndroid versions
+last 1 iOS versions
+maintained node versions
