+++ .npmrc
@@ -1,0 +1,3 @@
+//npm.pkg.github.com/:_authToken=${NPM_TOKEN}
+@jolution:registry=https://npm.pkg.github.com/
+
