+++ .nvmrc
@@ -1,0 +1,1 @@
+lts/*
