+++ components/config.js
