+++ components/content/translation.js
@@ -1,19 +1,19 @@
-{
-  "en": {
-    "COPY_GIT_COMMAND": "Copy git checkout command",
-    "SELECT_PROMPT": "Please select",
-    "COPY_BUTTON_TEXT": "Copy",
-    "TYPE": "Type",
-    "REQUIRED": "Required"
-  },
-  "de": {
+export default {
+  de: {
     "COPY_GIT_COMMAND": "Git checkout Befehl kopieren",
     "SELECT_PROMPT": "Bitte auswählen",
     "COPY_BUTTON_TEXT": "Kopieren",
     "TYPE": "Art",
     "REQUIRED": "Erforderlich"
   },
-  "fr": {
+  en: {
+    "COPY_GIT_COMMAND": "Copy git checkout command",
+    "SELECT_PROMPT": "Please select",
+    "COPY_BUTTON_TEXT": "Copy",
+    "TYPE": "Type",
+    "REQUIRED": "Required"
+  },
+  fr: {
     "COPY_GIT_COMMAND": "Copier la commande de validation git",
     "SELECT_PROMPT": "Veuillez sélectionner",
     "COPY_BUTTON_TEXT": "Copier",
